feat(editor): apply project icon to Windows build executable
